+++ app/src/main/kotlin/se/gustavkarlsson/skylight/android/services_impl/providers/KlausBrunnerDarknessProvider.kt
@@ -16,6 +16,6 @@ constructor() : DarknessProvider {
     override fun getDarkness(time: Instant, latitude: Double, longitude: Double): Darkness {
         val date = GregorianCalendar().apply { timeInMillis = time.toEpochMilli() }
         val azimuthZenithAngle = Grena3.calculateSolarPosition(date, latitude, longitude, 0.0)
-        return Darkness(azimuthZenithAngle.zenithAngle.toFloat())
+        return Darkness(azimuthZenithAngle.zenithAngle)
     }
 }
