+++ app/src/main/kotlin/se/gustavkarlsson/skylight/android/services_impl/providers/RetrofittedGeomagActivityProvider.kt
@@ -5,8 +5,8 @@ import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.debug
 import se.gustavkarlsson.aurora_notifier.common.service.KpIndexService
 import se.gustavkarlsson.skylight.android.R
-import se.gustavkarlsson.skylight.android.services.providers.GeomagActivityProvider
 import se.gustavkarlsson.skylight.android.entities.GeomagActivity
+import se.gustavkarlsson.skylight.android.services.providers.GeomagActivityProvider
 import se.gustavkarlsson.skylight.android.util.UserFriendlyException
 import java.io.IOException
 import javax.inject.Inject
@@ -26,7 +26,7 @@ constructor(
                 throw UserFriendlyException(R.string.error_could_not_determine_geomag_activity, response.errorBody()!!.string())
             }
             val kpIndex = response.body()!!
-            return GeomagActivity(kpIndex.value)
+            return GeomagActivity(kpIndex.value.toDouble())
         } catch (e: IOException) {
             throw UserFriendlyException(R.string.error_could_not_determine_geomag_activity, e)
         }
